Allow self-references in top-level package list
It now allows for package to use our own packages easily.

Though, having a recursive attribute set can be easier, setting our
package list this way is easier to maintain in the long-term since it is
a set-and-forget setup. The only disadvantage I can see is the need to
re-review what the hell the code is doing depending on my familiarity
for Nix in the future. ;p