hosts/ni: update config
For now, we're removing the backup media task just to make it easier to
install for non-local setups. It is pretty much the only thing that
requires the most secrets and I want to easily showcase my NixOS setup
or something like that. Enabling the task is still in the local changes,
just not to be committed.