config: remove foo-dogsquared's binary cache
It is pretty outdated and the CI didn't able to build the packages for
some time now (that I'm lazy to fix). Not to mention, some of them are
being upstreamed to nixpkgs so the need to use this is low. Also, it
causes the build to have some errors.