profiles/dev: remove extra Bash functions
It turns out fzf has some niceties with history and file selection which
is closer to how I imagine it. PRETTY NICE!

But it does override my Atuin configuration which doesn't have its shell
integration enabled. It does have some conflict with the Wezterm
shell integration. This is a problem for future me to deal with.