config: filter NixOS configuration
With images that are only exported as ISO (or even without one) are
included as part of NixOS configurations.

The idea with this attribute seems to be closer for deploying and
installing them with `nixos-install` which most of them aren't. This
might be reverted at some point. It's not clear to me yet.