workflows/mosey-branch: create desktop session files package
This should make it easier to manage the custom desktop files with the
usual toolchain for configuring and installing custom desktop files
(i.e., Meson). While this could be done with Nix, some parts of
installing files is easier with Meson.

Who knows, it might be reverted because my Nix-fu is pretty weak right
now.