docs: update project README
I've been using less and less Flatpak and more Nix-built packages. The
advantages of copying the Flatpak runtime is pretty nice but it faces
similar problems to simply migrating your home directory from one
installation to another such as version config mismatches and different
cache that might affect newer installations in a subtle way. It is often
better to have a new runtime directory and selectively back up all of
the required apps that you want to migrate.